media: smfc: add support for JPEG decompression
HWJPEG is capable of decompression of baseline, sequaltial DCT, Huffman
coded JPEG streams to an arbitrary image format that is supported by
HWJPEG. HWJPEG requires the address of SOS marker to start
decompression. Thus the driver should configure quantization tables,
Huffman tables and chroma subsampling factors to HWJPEG separately with
the JPEG stream address (address of SOS marker).
It is convenient to parse a JPEG header in userspace. But the driver
require the address of SOI marker to parse JPEG headers in the driver
because the driver does not convice the userspace to provide correct
information for decompression. Moreover, V4L2 does not consider about
JPEG decompression. It does not provide interface for passing Huffman
tables and quantization tables to the driver from userspace.

JPEG stream can be simply decompressed by passing the address of SOI.
If the given JPEG stream has error or is not supported, the userspace
will get an error from the driver.

Change-Id: I5322fc39b925d6cf06cb336c62bc465ee1b8e7f8
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>
(cherry picked from commit 667658e457960bc596b22218449059b86955b4be)
Signed-off-by: Giuliano Procida <gprocida@google.com>