arch/arm64: Add lazy preempt support
arm64 is missing support for PREEMPT_RT. The main feature which is
lacking is support for lazy preemption. The arch-specific entry code,
thread information structure definitions, and associated data tables
have to be extended to provide this support. Then the Kconfig file has
to be extended to indicate the support is available, and also to
indicate that support for full RT preemption is now available.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>
[Sultan Alsawaf: moved TIF_SYSCALL_TRACE to make _TIF_WORK_MASK bits
 contiguous, since _TIF_WORK_MASK requires it]
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>