media: smfc: use s/w timer if HWFC is enabled
If HWFC is enabled, the internal timer in the H/W is not available
because when the H/W finishes depends on the outside of SMFC itself.
Instead of the internal timer, S/W timer is used for the case of HWFC.
S/W timer is expired if the H/W is not completed in a second that is
quite huge amount of time.

Change-Id: I89dcb081db4e1e689d13bbbcf5cfaf05c17137c2
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>
(cherry picked from commit 3515a16906c00a7a3cdc8c3a174730738150dfa7)
Signed-off-by: Giuliano Procida <gprocida@google.com>