media: smfc: add support for the same H/W of different versions
SMFC that is also known as FIMP-JPEG has various H/W versions with
different capabilities. The HWJPEG of Exynos7420 is capable of
downscaling and cropping during decompression while that of Exynos8890
is not. HWJPEG of Exynos3450 does not have the decompression fuction.
Exynos8890, on the other hand, has the advanced functions such as the
hardware flow control(HWFC) and back-to-back compression.

The various version information is distinguished by different
compatible string and the detailed information about each version is
stored in the data field of of_device_id. The driver then uses the
version information in .probe().

Change-Id: Idf7432375d7a82cdbb83d81c65d06652503d5806
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>
(cherry picked from commit 4bc667b8060beb7387dc8119271e16bfd24a0a36)
Signed-off-by: Giuliano Procida <gprocida@google.com>