[COMMON] g2d: add priority setting
All contexts have priority. The requested task inherits
the priority of the context. Updating the priority of the context
always succeeds, but EBUSY could be returned if other tasks is running
or queued with a priority lower than its own.

If the priority of the context is lower than the priority of
the other context, the driver doesn't accept lower context's request,
It is called S/W preemption.

Change-Id: I82650f516ac019eb77f853a948b7f11cfe0578ae
Signed-off-by: hyesoo.yu <hyesoo.yu@samsung.com>
(cherry picked from commit 2e70fd6c89bc85eab1dff10728e7ff7bcda0f69e)
Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 165562344