[COMMON] g2d: cancel callback of acquire fence
If the user request has error parameter, the acquire
fence was put but the callback of fence is remained.
The new request on the same task structure registers
the callback to another fence, the list was corrupted.

It makes panic or hard lockup on system, so fix it.

Change-Id: I19ccbc06668c0d418fee675bf53188c6c37db3f9
Signed-off-by: hyesoo.yu <hyesoo.yu@samsung.com>
(cherry picked from commit 83092100e96177bc6cbd52165be5385dc48b5c97)
Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 165562344