google/debug: Add dbgcore directory in debugfs
Cherry-pick of 'e9cbaef' with conflict resolution & code adjustment.

The dbgcore directory in the debugfs is to provide helpful information
for the debug core.  It contains two files for now - version and
logdump.  The 'version' is to read debug core build version.

The logdump file is to read debug core's log messages from AP.  When
writing to the file, it'll send IPC_CMD_COPY_DEBUG_LOG message so that
debug core can write its log data to DRAM and when reading it'll read
contents from the DRAM region.

The format of the logdump file is like following:

  <seq-no>: <plugin>: <string>: <value>

For example, you can read the log as below:

  # cd /sys/kernel/debug/dbgcore

  ;; to update log from debug core
  # echo > logdump

  ;; read version string
  # cat version
  namhyung-8f73644-20191118163923

  ;; read the log
  # cat logdump
  0xd300: BCM:  rcvPipc:   0x200001
  0xd301: BCM: plugipc+:       0x47
  0xd302: BCM:  pd_info:    0x70001
  0xd303: BCM: plugipc-:        0x1
  0xd304: FRM:    Reply:        0x1
  0xd305: FRM:  irq_num:        0x1
  0xd306: BCM:  rcvPipc:   0x200001
  0xd307: BCM: plugipc+:        0x7
  ...

Bug: 156057245
Signed-off-by: Namhyung Kim <namhyung@google.com>
Change-Id: Ieabaf8bb2c50fba963433519a76adb22fd03abcc