drm: samsung: add display hibernation feature
Hibernation is an important feature from the point of view
power consumption.

If a panel with self refresh enabled is used, DPU hardware can be off
if there is no screen update while the LCD is on.

The KMS driver attempts to enter the hibernation if there is no screen
update for 3 TE irqs, and it escapes from the hibernaton if there is
new screen update.

Signed-off-by: Jiun Yu <jiun.yu@samsung.com>
(cherry picked from commit 074546bdd63ac3b2c885ce92d9957548b9412855)

Change-Id: I8811a11ad269c04abf03d641fb3dfe0e953fca37
Signed-off-by: Will McVicker <willmcvicker@google.com>