ufs: introduce cal that separates function and data
There has been a possibility that one of the projects
which use same or similar version with a specific version
of uic logic needs to add something but it's not applied
to other projects using the version of uic logic. That's
because even FUNCTIONS as well as DATA for uic configurations
has been separated. From now on, the FUNCTIONS need to
guarantee backward compatibility completely and if any change
happens, it should be broadcasted to all the projects using
this structure. DATA could change for respective projects.

To prevent from using incompatible combination of FUNCTIONS
and DATA completely, we have some definitions and
UFS_VS_MMIO_VER and UFS_CAL_IF_COMPAT_MMIO_VER should be the same and
UFS_CAL_IF_VER and UFS_CAL_TABLE_COMPAT_IF_VER should be so.
When a new project starts, you're used to start with
the latest cal data of the latest project
which uses same or similar uic logic.
If cal interface version, i.e. UFS_CAL_IF_VER increases for a reason
in another project, a build error will occur in the new project and
humans will recognize what happens.

Another big change is to use direct path, not uic command, to
configure uic attributes. The uic command requires sfr accesses
at least more than four times and even interrupt handling.
With this change, the only thing that exynos drivers has to do is
to let cal know what the base addresses for all the regions are,
except for delay function which can't be implemented only by software.

Let me introduce files for this structure:
ufs-vs-mmio.h: containing base address and macros to access sfrs
ufs-cal-if.c: functions to configure uic (a.k.a. cal function)
ufs-cal-if.h: declarations of public functions and their related stuffs
ufs-cal-[PROJECT].h: data (a.k.a. cal data)

Change-Id: Id2543dc052e3f609661a6a0af2a0eea687c119dc
Signed-off-by: Kiwoong Kim <kwmad.kim@samsung.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>