drm: samsung: cleanup bts win config
Currently BTS related win config is reset on every atomic commit, this
is unnecessary if some planes haven't changed. And could be problematic
if only a few planes are part of the atomic commit state. Fix by only
updating BTS win_config for planes that are in commit state and clear
unused windows.

Bug: 157778888
Change-Id: Id92eb6014c71a80289b80342ee71f22c0c4e1e3d
Signed-off-by: Adrian Salido <salidoa@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>