PT: Fix wait_event call for in-progress resize op
The condition that is used in the wait_event() call in
pt_resize_list_disable() appears to be backwards. The second parameter
of wait_event is defined as follows:

"The process is put to sleep (TASK_UNINTERRUPTIBLE) until the @condition
evaluates to true. The @condition is checked each time the waitqueue
@wq_head is woken up."

It appears as if the intend here was to wait for this particular pts to
*not* be or no longer be in progress.

Bug: 161812974
Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: Ic1f6b3101d2aaf28b425d207cd7395b1fb6f7d99
(cherry picked from commit 51f51fbb3a7ddcd19bd9697e729bafaaaddafd04)
Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 167676833