acpm: Support raw debug logs
This change adds support for raw ACPM debug logs. If a message is a raw
debug log, the kernel prints the raw bits instead of parsing the log as
an eight byte string and a four byte value. This will come in handy for
developing tools like A-B capture of DVFS requests.

A corresponding change was added to the ACPM firmware for inserting raw
bits in the debug log.

Bug: 151111906
Change-Id: Ied7426fb77a3d3c5b27b5b326e44dce9dcf58a2a
Signed-off-by: Jorge Troncoso <jatron@google.com>
Signed-off-by: Thierry Strudel <tstrudel@google.com>
(cherry picked from commit 627954a64cd2f2a1c36e9dc8659f3263b18bec97)